Hw 14 (#46)
* add basic auth guard

* add blog body validation

* add posts body validation

* add users body validation

* add HttpExceptionFilter & GlobalValidationPipe

* small refactor HttpExceptionFilter

* add rate limit to auth & create folder for local/jwt strategies

* add custom trim decorator

* add auth dto

* add email adapter & email manager

* change login return status

* set config

* set esLint config

* add registration route & test

* add registration-confirmation route & tests & delete passport strategies & refactor user schema

* refactor updateConfirmEmailStatus in users repo

* add resending email route & tests

* add login fn & tests

* add cookie parser

* fix login endpoint

* fix login endpoint #2

* fix login endpoint #2

* add fix

* fix confirm register endpoint

* fix

* fix users repo

* fix users pagination & user schema

* fix users pagination in query repo & user schema

* fix users

* add blog/post exist decorators

* fix blog exist decorator & add blog endpoint tests

* add post like fn

* refactor post like status fn & create/get comment by post

* add comments update, delete, update like status fn & tests

* fix blogId exist decorator

* fix create post by blog decorator validation

* add check if comment not found

+++ src/modules/comments/comments.service.ts
@@ -25,6 +25,10 @@ export class CommentsService {
   }
 
   async updateComment(commentId: string, dto: updateCommentDto) {
+    const comment = await this.commentsRepo.checkCommentById(commentId);
+
+    if (!comment) return false;
+
     return await this.commentsRepo.updateComment(commentId, dto.content);
   }
 
@@ -33,6 +37,10 @@ export class CommentsService {
   }
 
   async deleteComment(commentId: string) {
+    const comment = await this.commentsRepo.checkCommentById(commentId);
+
+    if (!comment) return false;
+
     return await this.commentsRepo.deleteComment(commentId);
   }
 }
